Move romeo-related code to publishers app, where it belongs
